Public --> External
onlyOwner functions by default cannot be called by the contract itself so should be set to external rather than public. The difference between the two is the following: public does not know if it will use calldata (cheaper) or memory (more expensive) and hence it needs more gas. External functions know they will only use calldata and so use less gas as they do not have to determine the location of the variables during execution